Use dot-notation to get settings
Lets you get settings by calling `Voyager::setting('group.key')`.
Also you can use `Voyager::setting('group')` to get all settings in a
group as an array.

To make use of this you have to update the keys of the settings by going
to the Settings-Page (and optionally group your settings) and save the
settings.